backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
sqlalchemy>=2.0.0
aiosqlite>=0.19.0
telethon>=1.34.0
python-dotenv>=1.0.0
pydantic>=2.0.0
cryptography>=41.0.0
